% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrAnnotation-data.R
\docType{data}
\name{chrAnnotation}
\alias{chrAnnotation}
\title{Sample CpG Annotation Dataset}
\format{
## `chrAnnotation`
A data frame with 240 rows and 52 columns:
\describe{
  \item{IlmnID, Name}{CpG names}
  \item{CHR}{Chromosome}
  \item{MAPINFO}{Position for CpG in chromosome}
  \item{UCSC_RefGene_Name}{Gene names}
  ...
}
}
\source{
<https://support.illumina.com/array/array_kits/infinium-methylationepic-beadchip-kit/downloads.html>
}
\usage{
data(chrAnnotation)
}
\description{
A subset of CpG annotation data from the Illumina.
}
\keyword{datasets}
