% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_extract.R
\name{cpg_extract}
\alias{cpg_extract}
\title{Extract subset of CpG sites}
\usage{
cpg_extract(
  db_obj,
  gene_list = FALSE,
  gene_col = FALSE,
  select_chr = FALSE,
  chr_col = FALSE,
  site_names = FALSE,
  row_indices = FALSE
)
}
\arguments{
\item{db_obj}{MethList created using \code{\link{create_methlist}}}

\item{gene_list}{Character string of gene names to select CpGs from. Default: `FALSE`.}

\item{gene_col}{Column for gene names in the CpG annotation data from `Methlist`.}

\item{select_chr}{Character string of chromosome number to select CpGs from. Default: `FALSE`.}

\item{chr_col}{Column for chromosome in `cpg_annot` of the Methlist. Default: `FALSE`}

\item{site_names}{Character string for CpG names}

\item{row_indices}{Numeric vector indicating the row indices to extract CpG data}
}
\value{
Subset of CpG data along with their annotation based on selection criteria
}
\description{
cpg_extract extracts subset of CpG methylation data based on argument provided
}
\examples{
library(tidyverse)
library(arrow)
data(phenoData)
data(chrAnnotation)
data(ewas_bmi)
wdir <- getwd()
path <- paste0(wdir,'/data/Parquet_Directory')

# Create MethList
mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')

# Extract CpG data based on chromosome
chr <- as.character(seq(1:5))
extr_chr <- cpg_extract(mlist,select_chr=chr,chr_col='CHR')
head(extr_chr[[1]])[1:5,1:5]
head(extr_chr[[2]])
# Based on row index
extr_row <- cpg_extract(mlist,row_indices = 1:100)
}
