% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_methlist.R
\name{create_methlist}
\alias{create_methlist}
\title{Create Methlist}
\usage{
create_methlist(
  db_path,
  cpg_col_db,
  subject_annot,
  subject_col_keep = "all",
  subject_id,
  cpg_annot,
  cpg_col_annot,
  cpgAnnot_col_keep = "all",
  gene_col_name = FALSE,
  chr_col_name = FALSE
)
}
\arguments{
\item{db_path}{Filepath to a methylation parquet database. Should contain a column named 'CpG'
for CpG site names}

\item{cpg_col_db}{Column for CpG site names in parquet data, which will be named as `CpG` in
CpG data inside meth list}

\item{subject_annot}{Subject annotation file}

\item{subject_col_keep}{Numeric vector containing columns to include in the subject annotation
file. Default value: `all`}

\item{subject_id}{Column for subject IDs in `subject_annot`, which will be named as `sample_id`
in `subject_annot` inside meth list}

\item{cpg_annot}{CpG annotation file}

\item{cpg_col_annot}{Column for CpG site names in `cpg_annot`}

\item{cpgAnnot_col_keep}{Numeric vector containing columns to include in the CpG annotation
file. Default value: `all`}

\item{gene_col_name}{Name of column containing gene names in `cpg_annot`. If povided, this column
will be named as `Gene` in `cpg_annot` of the meth list. Default value: `FALSE`}

\item{chr_col_name}{Name of column containing chromosome number in `cpg_annot`. If povided,
this column will be named as `CHR` in `cpg_annot` of the meth list. Default value: `FALSE`}
}
\value{
A `MethList` object which contains:
`db`: open connection to a methylation parquet databse
`subject_annot`: subject annotation file
`cpg_annot`: CpG annotation file
}
\description{
Create Methlist
}
\examples{
library(tidyverse)
library(arrow)
data(MethData)
data(phenoData)
data(chrAnnotation)
wdir <- getwd()
path <- paste0(wdir,'/data/Parquet_Directory')

# Create Parquet data in 'path'
MethData \%>\% group_by(CHR) \%>\% arrow::write_dataset(path,format = "parquet")

# Create MethList object
mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')
names(mlist)
}
