% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_methlist_S3.R
\name{create_methlist_S3}
\alias{create_methlist_S3}
\title{Create MethList from Amazon Web Services S3 bucket}
\usage{
create_methlist_S3(
  db_s3_path,
  subject_annot = FALSE,
  subject_path = FALSE,
  cpg_annot = FALSE,
  cpg_path = FALSE,
  subject_readFun = FALSE,
  cpg_readFun = FALSE,
  access_key = FALSE,
  secret_access_key = FALSE,
  default_region = FALSE,
  cpg_col_db,
  subject_col_keep = "all",
  subject_id,
  cpg_col_annot,
  cpgAnnot_col_keep = "all",
  gene_col_name = FALSE,
  chr_col_name = FALSE
)
}
\arguments{
\item{db_s3_path}{Filepath to the methylation parquet database in AWS_S3 bucket. Should contain a column
named 'CpG' for CpG site names}

\item{subject_annot}{Subject annotation file}

\item{subject_path}{Filepath to the subject annotation file in AWS_S3 bucket if not provided as
`subject_annot`. Default value: `FALSE`}

\item{cpg_annot}{CpG annotation file}

\item{cpg_path}{Filepath to the CpG annotation file in AWS_S3 bucket if not provided as
`cpg_annot`. Default value: `FALSE`}

\item{subject_readFun}{Function to read the subject annotation file in AWS_S3 bucket.}

\item{cpg_readFun}{Function to read the cpg annotation file in AWS_S3 bucket.}

\item{access_key}{access key for the AWS_S3 bucket, if applicable.}

\item{secret_access_key}{secret access key for the AWS_S3 bucket, if applicable.}

\item{default_region}{AWS default region, if applicable.}

\item{cpg_col_db}{Column for CpG site names in parquet data, which will be named as `CpG` in
CpG data inside meth list}

\item{subject_col_keep}{Numeric vector containing columns to include in the subject annotation
file. Default value: `all`}

\item{subject_id}{Column for subject IDs in `subject_annot`, which will be named as `sample_id`
in `subject_annot` inside meth list}

\item{cpg_col_annot}{Column for CpG site names in `cpg_annot`}

\item{cpgAnnot_col_keep}{Numeric vector containing columns to include in the CpG annotation
file. Default value: `all`}

\item{gene_col_name}{Name of column containing gene names in `cpg_annot`. If povided, this column
will be named as `Gene` in `cpg_annot` of the meth list. Default value: `FALSE`}

\item{chr_col_name}{Name of column containing chromosome number in `cpg_annot`. If povided,
this column will be named as `CHR` in `cpg_annot` of the meth list. Default value: `FALSE`}
}
\value{
A `MethList` object which contains:
`db`: open connection to a methylation parquet databse in AWS_S3
`subject_annot`: subject annotation file
`cpg_annot`: CpG annotation file
}
\description{
create_methlist_S3 creates the MethList object linking parquet database stored in
Amazon Web Services(AWS) Simple Storage Service (S3) bucket.
}
\examples{
# Example code only

# mlist=create_methlist_S3(db_s3_path = "path to methylation Parquet data in AWS_S3",
# subject_path="path to subject annotation file in AWS_S3",
# cpg_path="path to CpG annotation file in AWS_S3",
# access_key = "access key to S3 bucket", secret_access_key = "secret access key to S3 bucket",
# default_region = "us-east-1", cpgAnnot_col_keep=c(1:2,12:13,16),
# subject_readFun = load, cpg_readFun = read.csv, cpg_col_db='CpG',
# subject_id='sample_id', cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')
}
