% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_meth_score.R
\name{dev_meth_score}
\alias{dev_meth_score}
\title{Calculation of Methylation Risk Score (MRS)}
\usage{
dev_meth_score(
  db_obj,
  meth_effect_data,
  cpg_col,
  scale_score = FALSE,
  effect_col,
  divide_by_site_n = FALSE,
  NAs_to_zero = FALSE
)
}
\arguments{
\item{db_obj}{MethList created using \code{\link{create_methlist}}}

\item{meth_effect_data}{Phenotype data containing CpG sites and effect}

\item{cpg_col}{Column for CpG site names in `meth_effect_data`}

\item{scale_score}{Scale score using \code{\link[base]{scale}} function. Default: `FALSE`}

\item{effect_col}{Column for effect in `meth_effect_data`}

\item{divide_by_site_n}{Divide scores by number of methylation sites Default: `FALSE`}

\item{NAs_to_zero}{Convert any missing values in methylation data (inside Methlist) to 0. Default: `FALSE`}
}
\value{
A matrix of methylation risk scores associated to the phenotype for each subject
}
\description{
dev_meth_score calculates Methylation Risk Score as a weighted sum of an individual’s
DNA methylation values and pre-calculated weights associated with a specific phenotype.
}
\examples{
library(tidyverse)
library(arrow)
data(phenoData)
data(chrAnnotation)
data(ewas_bmi)
data(MethData)
wdir <- getwd()
path <- paste0(wdir,'/Parquet_Directory')

# Create Parquet data in 'path' and MethList
MethData \%>\% group_by(CHR) \%>\% arrow::write_dataset(path,format = "parquet")

mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')

# Calculation of MRS associated with BMI
mrs_bmi <- dev_meth_score(db_obj=mlist,cpg_col='CpG',meth_effect_data = ewas_bmi,effect_col='Effect')
head(mrs_bmi)
class(mrs_bmi)
unlink(path,recursive=TRUE)
}
