% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas_meth_exposure.R
\name{ewas_meth_exposure}
\alias{ewas_meth_exposure}
\title{Association analysis with methylation as exposure}
\usage{
ewas_meth_exposure(
  db_obj,
  m_null,
  select_sites = "full",
  select_chr = FALSE,
  gene_list = FALSE,
  gene_col = FALSE,
  NAs_to_zero = FALSE,
  out_position = FALSE,
  block_size = 50000,
  parallel = FALSE
)
}
\arguments{
\item{db_obj}{MethList created using \code{\link{create_methlist}}.}

\item{m_null}{Null model created by function \code{\link{NullModel}}}

\item{select_sites}{Character string consisting of CpG(s) to be tested. Default: `full` (all CpGs).
Only one selection for CpG is allowed, `sites`, `chromosome`, or `gene`.}

\item{select_chr}{Character string of chromosome number to select CpGs from. Default: `FALSE`.}

\item{gene_list}{Character string of gene names to select CpGs from. Default: `FALSE`.}

\item{gene_col}{Column for gene names in the CpG annotation data from `Methlist`.}

\item{NAs_to_zero}{Convert missing values to 0 in methylation data. Default: `FALSE`.}

\item{out_position}{Column name for CpG position in chromosome from `cpg_annot` of `MethList`.
If given, output will contain these columns and chromosome number.}

\item{block_size}{Integer to specify number of CpGs in each iteration block. Default: 50,000.}

\item{parallel}{Number of cores used for parallel processing (forking). For details please see
\code{\link[future]{plan}}.  Default: FALSE (sequential processes).}
}
\value{
A data frame with coefficient estimates for CpGs as exposure and testing statistics.
}
\description{
ewas_meth_exposure computes association analysis for each CpG site using the null model
fitted by function \code{\link{NullModel}}
}
\examples{
library(tidyverse)
library(arrow)

wdir <- getwd()
methpath <- system.file('extdata','MethData.csv',package='MethParquet')
data(phenoData)
data(chrAnnotation)

# Create Parquet data and MethList
path <- paste0(wdir,'/Parquet_Directory')
write_parquet_meth(data_path=methpath,format='csv',group_by='CHR',parquet_path = path)

mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')

# Fit linear null model (default)
m.null=NullModel(db_obj=mlist,trait='bmi',covariates_string = c('age','sex'))

# Testing association between each CpG and trait using fitted null model
lm_v=ewas_meth_exposure(db_obj=mlist,m.null,select_sites = 'full',select_chr = FALSE,
gene_list=FALSE,out_position='MAPINFO')
head(lm_v)
unlink(path,recursive=TRUE)
}
