% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_ewas.R
\name{flex_ewas}
\alias{flex_ewas}
\title{Flexible implementation of association functions on methylation data}
\usage{
flex_ewas(
  db_obj,
  fun,
  phen = NULL,
  select_sites = "full",
  select_chr = FALSE,
  gene_list = FALSE,
  gene_col = FALSE,
  out_position = FALSE,
  block_size = 50000,
  NAs_to_zero = FALSE
)
}
\arguments{
\item{db_obj}{MethList created using \code{\link{create_methlist}}.}

\item{fun}{A regression function applied to each CpG site}

\item{phen}{Phenotype data based on which the function (`fun`) is built. Default: `NULL` is the subject annotation from the `db_obj`}

\item{select_sites}{Character string consisting of CpG(s) to be tested. Default: `full` (all CpGs).}

\item{select_chr}{Character string of chromosome number to select CpGs from. Default: `FALSE`.}

\item{gene_list}{Character string of gene names to select CpGs from. Default: `FALSE`.}

\item{gene_col}{Column for gene names in the CpG annotation data from `Methlist`.}

\item{out_position}{Column names for CpG position in chromosome from `cpg_annot` of `MethList`.
If given, output will contain these columns and chromosome number.}

\item{block_size}{Integer to specify number of CpGs in each iteration block. Default: 50,000.}

\item{NAs_to_zero}{Convert missing values to 0 in methylation data. Default: `FALSE`.}
}
\value{
A data frame with results obtained using `fun`.
}
\description{
`flex_ewas` takes another regression function as input to test association between
each CpG site and the phenotype defined in the regression model
}
\examples{
library(tidyverse)
library(arrow)
data(phenoData)
data(chrAnnotation)
wdir <- getwd()
path <- paste0(wdir,'/data/Parquet_Directory')

# Create MethList
mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')

# Create a robust linear regression function with methylation as exposure
rlm_function <- function(x) {
mod <- base::suppressWarnings(MASS::rlm(as.formula(paste('height', "~x+", 'age')), data = phenoData))
return(summary(mod)$coefficients[2,])
}

# Run flex_ewas
flex_rlm <- flex_ewas(mlist,rlm_function,out_position = c('MAPINFO','Gene'))
head(flex_rlm)
}
