% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethData-data.R
\docType{data}
\name{MethData}
\alias{MethData}
\title{Sample Methylation Dataset}
\format{
## `MethData`
A data frame with 240 rows and 102 columns:
\describe{
  \item{CpG}{CpG names}
  \item{CHR}{Chromosome}
  \item{GSA###, TCGA-####}{Subject names}
}
}
\source{
<https://ngdc.cncb.ac.cn/ewas/datahub/download>
}
\usage{
data(MethData)
}
\description{
A subset of BMI methylation data from the EWAS Data Hub.
}
\keyword{datasets}
