% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethParquet-package.R
\docType{package}
\name{MethParquet-package}
\alias{MethParquet}
\alias{MethParquet-package}
\title{MethParquet: Manipulate and analyze methylation data using Apache Parquet}
\description{
MethParquet leverages the columnar format Parquet for efficient DNA methylation (DNAm) data analysis. Epigenome-wide association analysis (EWAS), methylation risk score (MRS) calculation and testing, as well as CpG subset extraction are streamlined via the Methlist object containing connection to methylation Parquet directory along with CpG and subject annotation data.
}
\author{
\strong{Maintainer}: Tamar Sofer \email{tsofer@bidmc.harvard.edu}

Authors:
\itemize{
  \item Ziqing Wang \email{zwang14@bidmc.harvard.edu}
  \item Michael Cassidy
  \item Danielle Clarkson-Townsend \email{dclarkson-townsend@bwh.harvard.edu}
}

}
\keyword{internal}
