% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoData-data.R
\docType{data}
\name{phenoData}
\alias{phenoData}
\title{Sample Subject Annotation Dataset}
\format{
## `phenoData`
A data frame with 100 rows and 38 columns:
\describe{
  \item{sample_id}{Subject IDs}
  \item{age}{Age}
  ...
}
}
\source{
<https://ngdc.cncb.ac.cn/ewas/datahub/download>
}
\usage{
data(phenoData)
}
\description{
A subset of subject annotation data for BMI methylation from the EWAS Data Hub.
}
\keyword{datasets}
