% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MethPheno.R
\name{plot_MethPheno}
\alias{plot_MethPheno}
\title{Plot methylation against phenotype}
\usage{
plot_MethPheno(db_obj, sites, trait, output_plot = F, width = 9, height = 8)
}
\arguments{
\item{db_obj}{MethList created using \code{\link{create_methlist}}}

\item{sites}{Character string for CpG names}

\item{trait}{Exposure phenotype in `subject_annot` from `MethList`.}

\item{output_plot}{File path to save the plot. Default: FALSE.}

\item{width}{Width in inch for the saved output plot when given path (`output_plot`). Default: 9}

\item{height}{Height in inch for the saved output plot when given path (`output_plot`). Default: 8}
}
\value{
A plot with methylation beta values on y axis against the phenotype on x axis
}
\description{
Create a plot of methylation against phenotype for each CpG site
}
\examples{
library(tidyverse)
library(arrow)
library(ggplot2)
data(phenoData)
data(chrAnnotation)

wdir <- getwd()
methpath <- paste0(wdir,'/inst/extdata/MethData.csv')

# Create Parquet data and MethList
path <- paste0(wdir,'/Parquet_Directory')
write_parquet_meth(data_path=methpath,format='csv',group_by='CHR',parquet_path = path)

mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')

# Run linear model and get top results
ewas_lm <- lm_ewas_outcome(db_obj=mlist,trait='bmi',select_sites='full',select_chr=FALSE,
covariates_string = c('age','sex'),out_position='MAPINFO',block_size=100000)
sites<-ewas_lm$CpG[ewas_lm$fdr_bh<0.05][1:3]

# Plot top results
plot_MethPheno(mlist,sites=sites,trait='bmi',output_plot = FALSE)
unlink(path,recursive=TRUE)
}
