% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm_ewas_outcome.R
\name{rlm_ewas_outcome}
\alias{rlm_ewas_outcome}
\title{Association analysis with robust linear regression}
\usage{
rlm_ewas_outcome(
  db_obj,
  trait,
  select_sites = "full",
  select_chr = FALSE,
  gene_list = FALSE,
  gene_col = FALSE,
  covariates_string,
  NAs_to_zero = FALSE,
  out_position = FALSE,
  block_size = 50000
)
}
\arguments{
\item{db_obj}{MethList created using \code{\link{create_methlist}}.}

\item{trait}{Exposure phenotype in `subject_annot` from `MethList`.}

\item{select_sites}{Character string consisting of CpG(s) to be tested. Default: `full` (all CpGs).
Only one selection for CpG is allowed, `sites`, `chromosome`, or `gene`.}

\item{select_chr}{Character string of chromosome number to select CpGs from. Default: `FALSE`.}

\item{gene_list}{Character string of gene names to select CpGs from. Default: `FALSE`.}

\item{gene_col}{Column for gene names in the CpG annotation data from `Methlist`.}

\item{covariates_string}{Columns for covariates to adjust in model from `subject_annot` of `MethList`.}

\item{NAs_to_zero}{Convert missing values to 0 in methylation data. Default: `FALSE`.}

\item{out_position}{Column name for CpG position in chromosome from `cpg_annot` of `MethList`.
If given, output will contain these columns and chromosome number.}

\item{block_size}{Integer to specify number of CpGs in each iteration block. Default: 50,000.}
}
\value{
A data frame with coefficient estimates for `trait` and testing statistics.
}
\description{
rlm_ewas_outcome performs association analysis with robust linear regression
using methylation as the outcome.
}
\examples{
library(tidyverse)
library(arrow)
library(limma)
data(MethData)
data(phenoData)
data(chrAnnotation)
wdir <- getwd()
path <- paste0(wdir,'/Parquet_Directory')

# Create Parquet data in 'path' and MethList
MethData \%>\% group_by(CHR) \%>\% arrow::write_dataset(path,format = "parquet")

mlist <- create_methlist(db_path = path,cpg_col_db='CpG',subject_annot = phenoData,
subject_col_keep='all',cpgAnnot_col_keep=c(1:2,12:13,16),cpg_annot = chrAnnotation,
subject_id='sample_id',cpg_col_annot='Name', gene_col_name = 'UCSC_RefGene_Name')

ewas_rlm <- rlm_ewas_outcome(db_obj=mlist,trait='disease',select_sites=FALSE,gene_list=FALSE,
select_chr=c('1','2','21'),covariates_string = c('age','sex'),out_position='MAPINFO')
head(ewas_rlm)
unlink(path,recursive=TRUE)
}
