% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_parquet_meth.R
\name{write_parquet_meth}
\alias{write_parquet_meth}
\title{Write Parquet database from processed methylation data}
\usage{
write_parquet_meth(
  data_path,
  format = c("csv", "txt", "tsv"),
  chunk_size = 0,
  chr_col = "CHR",
  group_by = NULL,
  parquet_path
)
}
\arguments{
\item{data_path}{Path to methylation data.}

\item{format}{A string identifier of the file format, can be one of the following: ('csv','txt','tsv').}

\item{chunk_size}{Maximum number of rows per file in Parquet database. A value larger than 0 indicates
the number of rows of each single Parquet file. Default: 0L.}

\item{chr_col}{Character string indicating the column name for the chromosome number. Default: 'CHR'}

\item{group_by}{The variable (column name) of the methylation data to partition the Parquet data (as path segments).}

\item{parquet_path}{Path to write the Parquet directory}
}
\value{
A Parquet directory for the methylation data
}
\description{
write_parquet_meth writes Parquet database from methyltion data without loading it into
local machine. For a seamless MethParquet workflow, this function requires that the methylation data
contains a column representing chromosome number.
Currently supports comma-separated values (CSV), delimited text file (txt) and tab-separated values (TSV) files.
For more details please refer to \code{\link[arrow]{write_dataset}}.
}
\examples{
library(arrow)
library(tidyverse)
wdir <- getwd()
methpath <- system.file('extdata','MethData.csv',package='MethParquet')
path <- paste0(wdir,'/Parquet_Directory')
write_parquet_meth(data_path=methpath,format='csv',group_by='CHR',parquet_path = path)
list.files(path)
unlink(path,recursive=TRUE)
}
